plant-disease-inference/requirements.txt
tensorflow-cpu==2.12.0
numpy
pillow
